% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCDA.R
\docType{package}
\name{BCDA}
\alias{BCDA}
\alias{BCDA-package}
\title{Bayesian analysis of categorical data}
\description{
Bayesian analysis of categorical data
}
\references{
Agresti, A. and Hitchcock, D. B. (2005). Bayesian inference for categorical
  data analysis. \emph{Statistical Methods & Applications}, \strong{14},
  297-330. doi:10.1007/s10260-005-0121-y
}

