% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom.R
\name{beta_binom}
\alias{beta_binom}
\title{Fit a Binomial model with Beta priors}
\usage{
beta_binom(
  x,
  n = NULL,
  prior = c(a = 0.5, b = 0.5, c = 0.5, d = 0.5),
  n_sims = 10000
)
}
\arguments{
\item{x}{A numeric vector of success counts OR a 2x2 table or matrix.}

\item{n}{A numeric vector of totals. (Optional if x is a table or matrix.)}

\item{prior}{Parameters for independent beta priors. See \strong{Details}.}

\item{n_sims}{Number of draws (simulations) to make from the posterior
distributions.}
}
\value{
An S3 object of class "beta_binomial_fit". It has the following
  properties: posterior_simulations, posterior_parameters, successes, totals
}
\description{
Fits a Beta-Binomial model to a 2x2 table/matrix of two groups
  and two outcomes/features.
}
\details{
The model assumes
  \deqn{x_i ~ Binomial(n_i, p_i); p_1 ~ Beta(a, b); p_2 ~ Beta(c, d)}
  for \eqn{i = 1, 2}. The default is \code{a = b = c = d = 1/2} which
  corresponds to a Jeffreys prior on \eqn{p_1} and \eqn{p_2}.
}
\examples{
fake_data <- matrix(c(200, 150, 250, 300), nrow = 2, byrow = TRUE)
colnames(fake_data) <- c('Safe' ,'Dangerous')
rownames(fake_data) <- c('Animals', 'Plants')
beta_binom(fake_data)

beta_binom(x = c(200, 250), n = c(350, 550))
}
