% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_multinom.R
\name{est_multinom}
\alias{est_multinom}
\title{Estimate multinomial cell probabilities in a 2x2 table}
\usage{
est_multinom(x, n = NULL, prior = NULL, groups = c("test", "control"))
}
\arguments{
\item{x}{A numeric vector of success counts OR a 2x2 table.}

\item{n}{A numeric vector of totals.}

\item{prior}{A matrix of hyper-parameters.}
}
\description{
Simulatenously estimates of multinomial cell probabilities.
}
\details{
The model assumes no margins are fixed. Two random binary outcomes
  are measured on N individuals, resulting in four possible combined
  outcomes with probabilities \eqn{p_{11}, p_{12}, p_{21}, p_{22}}.
}
\examples{
data <- matrix(c(200, 150, 250, 300), nrow = 2, byrow = TRUE)
colnames(data) <- c('Safe' ,'Dangerous')
rownames(data) <- c('Animals', 'Plants')
est_multinom(data)

est_multinom(x = data[, 'Safe'], n = margin.table(data, 1))

hyperparams <- matrix(c(0.25, 0.2, 0.15, 0.4), nrow = 2, byrow = TRUE)
est_multinom(data, prior = hyperparams)

}
\references{
Fienberg, S.E. and Holland, P.W. (1973) Simultaneous estimation
  of multinomial cell probabilities. \emph{Journal of the American Statistical
  Association}, \strong{68}, 683-691. doi:10.1080/01621459.1973.10481405
}

