% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom.R
\name{glance.beta_binomial_fit}
\alias{glance.beta_binomial_fit}
\title{Glance at a fitted Beta-Binomial model.}
\usage{
\method{glance}{beta_binomial_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "beta_binomial_fit".}

\item{...}{Arguments to pass forward to [present_bbfit()]}
}
\description{
Constructs a single row summary "glance" of a Beta-Binomial model fit.
}
\examples{
\dontrun{
glance(beta_binom(x = c(200, 250), n = c(350, 550)))
}
}
