% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_indepen.R
\name{interpret_bf}
\alias{interpret_bf}
\title{Interpret Bayes Factor}
\usage{
interpret_bf(bf, interpreter = c("Kass and Raftery", "Harold Jeffreys"))
}
\arguments{
\item{bf}{A Bayes factor computed using \code{LearnBayes::ctable()}, for
example.}

\item{interpreter}{A character vector indicating whether to interpret
\code{bf} through the lens of Kass and Raftery or Harold Jeffreys.}
}
\examples{
\dontrun{
data <- matrix(c(200, 150, 250, 300), nrow = 2, byrow = TRUE)
colnames(data) <- c('Safe' ,'Dangerous')
rownames(data) <- c('Animals', 'Plants')

bf <- LearnBayes::ctable(data, matrix(rep(1, 4), 2))
interpret_bf(bf, interpreter = 'Kass and Raftery')
}
}

