% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_linear.R
\name{log_linear}
\alias{log_linear}
\title{Analyze a 2x2 table with a log-linear model.}
\usage{
log_linear(data, dim_names, n_sims = 1100, n_burn = 100, n_thin = 1)
}
\arguments{
\item{data}{A two-way table of counts.}

\item{dim_names}{Names of the variables.}

\item{n_sims}{Number of simulations, used in Bayes factor computation via
importance sampling.}

\item{n_burn}{Number of burn-in simulations to throw away.}

\item{n_thin}{Amount of thinning to use.}
}
\description{
Fits a Bayesian Generalized Linear Model to the contingency
  table.
}
\examples{
data <- matrix(c(200, 150, 250, 300), nrow = 2, byrow = TRUE)
colnames(data) <- c('Safe' ,'Dangerous')
rownames(data) <- c('Animals', 'Plants')

fit <- log_linear(data, c('kingdom', 'safety'))
conting::summary.bcct(fit)
}

