% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom.R
\name{plot.beta_binomial_fit}
\alias{plot.beta_binomial_fit}
\title{Visualize posterior draws from a fitted Beta-Binomial model}
\usage{
\method{plot}{beta_binomial_fit}(x, interval_type = c("quantile", "HPD"), ...)
}
\arguments{
\item{x}{An object of class "beta_binomial_fit".}

\item{interval_type}{Method for computing intervals ("quantile" or "HPD").}
}
\description{
Plots the point estimates and 95\% and 80\% credible intervals
  for parameters, including relative risk and odds ratio.
}

