% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom.R
\name{present_bbfit}
\alias{present_bbfit}
\title{Present the posterior results of the Beta-Binomial model}
\usage{
present_bbfit(object, conf_interval = TRUE, conf_level = 0.95,
  interval_type = c("quantile", "HPD"), ...)
}
\arguments{
\item{object}{An object of class "beta_binomial_fit" or a list (named or
unnamed) of "beta_binomial_fit" objects created by \code{beta_binom()}.}

\item{conf_interval}{A logical flag indicating whether to include Bayesian
confidence intervals in the generated table.}

\item{conf_level}{Probability level for credible intervals. 95\% by default.}

\item{interval_type}{Method for computing intervals ("quantile" or "HPD").}

\item{...}{Arguments to forward to \code{knitr::kable} (e.g. \code{format}).}
}
\value{
A character vector formatted as Markdown, HTML, or LaTeX. The table
  has the following columns:
  \describe{
    \item{Group 1}{The total count of trials in group 1.}
    \item{Group 2}{The total count of trials in group 2.}
    \item{Pr(Success) in Group 1}{Point estimate of the probability of success in group 1.}
    \item{Pr(Success) in Group 2}{Point estimate of the probability of success in group 2.}
    \item{Difference}{Pr(Success) in Group 1 - Pr(Success) in Group 2}
    \item{Relative Risk}{How likely trials in group 1 are to succeed relative to group 2.}
    \item{Odds Ratio}{The ratio of the odds of success in group 1 vs odds of success in group 2.}
  }
  The estimated quantities have credible intervals by default but these can
  be turned off with the \code{conf_interval} argument.
}
\description{
Outputs a nicely-formatted table suitable for presentations and
  reports. Especially useful for combining multiple results into a single
  summary table.
}
\examples{
\dontrun{
data <- matrix(c(200, 150, 250, 300), nrow = 2, byrow = TRUE)
colnames(data) <- c('Safe' ,'Dangerous')
rownames(data) <- c('Animals', 'Plants')

fit <- beta_binom(data)
present_bbfit(fit) # uses getOption("digits")
present_bbfit(fit, digits = 2)
present_bbfit(fit, conf_interval = FALSE, digits = 3)
present_bbfit(fit, conf_level = 0.8, interval_type = "HPD", digits = 2)

fit_2 <- update(fit, x = c(8, 4), n = c(10, 50))
fit_3 <- update(fit_2, x = c(1, 20), n = c(15, 40))
fit_4 <- update(fit_3, x = c(20, 13), n = c(80, 45))
present_bbfit(list("Day 1" = fit, "Day 2" = fit_2, "Day 3" = fit_3, "Day 4" = fit_4), digits = 2)
}
}

