% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_indepen.R
\name{test_independence}
\alias{test_independence}
\title{Test of independence for contigency tables using Bayes factor}
\usage{
test_independence(x, n = NULL, interpreter = c("Kass and Raftery",
  "Harold Jeffreys"))
}
\arguments{
\item{x}{A numeric vector of success counts OR a 2x2 table or matrix.}

\item{n}{A numeric vector of totals. (Optional if x is a table or matrix.)}

\item{interpreter}{A character vector indicating whether to interpret
\code{bf} through the lens of Kass and Raftery or Harold Jeffreys.}
}
\examples{
\dontrun{
data <- matrix(c(200, 150, 250, 300), nrow = 2, byrow = TRUE)
colnames(data) <- c('Safe' ,'Dangerous')
rownames(data) <- c('Animals', 'Plants')

test_independence(data)
test_independence(data, "Harold Jeffreys")
}

}
\references{
Albert, J. (2009) Bayesian Computation with R.
Springer-Verlag New York, New York. doi:10.1007/978-0-387-92298-0

Kass, R.E. and Raftery, A.E. (1995) Bayes Factors. \emph{Journal of
  the American Statistical Association}, \strong{90}, 773-795.
}
\seealso{
Irony, T.Z. and Pereira, C.A.B. (1986) Exact tests for equality
  of two proportions: fisher v. bayes. \emph{Journal of Statistical
  Computation and Simulation}, \strong{25}, 93-114.
  doi:10.1080/00949658608810926

Altham, P.M.E. (1963) Exact Bayesian analysis of a 2x2 contingency table,
  and Fisher's "exact" significance test. \emph{Journal of the Royal
  Statistical Society}, Series B, Methodological, \strong{31}, 261-269.

Altham, P.M.E. (1971) The analysis of matched proportions. \emph{Biometrika},
\strong{58}, 561-576.
}

