% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/beta_binom.R
\name{tidy}
\alias{tidy}
\alias{tidy.beta_binomial_fit}
\title{Tidy the model into a summary data.frame}
\usage{
\method{tidy}{beta_binomial_fit}(x, conf_level = 0.95,
  interval_type = c("quantile", "HPD"), ...)
}
\arguments{
\item{x}{An object of class "beta_binomial_fit".}

\item{conf_level}{Probability level for credible intervals. 95\% by default.}

\item{interval_type}{Method for computing intervals ("quantile" or "HPD").}
}
\value{
A data frame consistent with the output of David Robinson's
  \code{\link[broom]{tidyMCMC}} tidying method:
  \describe{
    \item{term}{Name of the parameter.}
    \item{estimate}{Point estimate (mean of posterior draws).}
    \item{std.error}{Standard error.}
    \item{conf.low}{Credible interval lower bound.}
    \item{conf.high}{Credible interval upper bound.}
  }
}
\description{
Summarize posterior draws from a fitted Beta-Binom model in a
  tidy form. Outputs a data frame of point estimates and Bayesian confidence
  intervals for group proportions, proportion difference, relative risk, and
  odds ratio.
}
\examples{
\dontrun{
tidy(beta_binom(x = c(200, 250), n = c(350, 550)))
}
}

