% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_table}
\alias{make_table}
\alias{flip_cols}
\alias{flip_rows}
\title{Utilities for tallied data frames}
\usage{
make_table(df)

flip_cols(x)

flip_rows(x)
}
\arguments{
\item{df}{A data.frame where the first column is the group column and the
rest are counts of outcomes.}

\item{x}{A 2-by-2 table produced by \code{make_table()}, \code{table()}, or
\code{xtabs()}.}
}
\value{
A 2-by-2 table.
}
\description{
Useful when piping a grouped data frame to \code{dplyr::tally()}
  and then converting that into a contingency table before piping it into
  \code{beta_binom}, or flipping rows and columns in between those steps.
  Works for data frames grouped by a single variable only, to create tables
  with more than 2 dimensions, use \code{xtabs}.
}
