% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom.R
\name{update.beta_binomial_fit}
\alias{update.beta_binomial_fit}
\title{Update the Beta-Binomial model posterior with new data}
\usage{
\method{update}{beta_binomial_fit}(object, x, n = NULL, n_sims = 1000)
}
\arguments{
\item{object}{An object of class "beta_binomial_fit".}

\item{x}{A numeric vector of success counts OR a 2x2 table or matrix.}

\item{n}{A numeric vector of totals. (Optional if \code{x} is a table or
matrix.)}

\item{sims}{Number of draws (simulations) to make from the posterior
distributions.}
}
\value{
An S3 object of class "beta_binomial_fit".
}
\description{
Uses the previously computed Beta posterior as prior to compute
  a new Beta posterior, which can subsequently be updated as even more data
  becomes available.
}
\examples{
\dontrun{
data <- matrix(c(200, 150, 250, 300), nrow = 2, byrow = TRUE)
colnames(data) <- c('Safe' ,'Dangerous')
rownames(data) <- c('Animals', 'Plants')

fit <- beta_binom(data)
fit <- update(fit, x = c(100, 200), n = c(400, 600))
}
}

